+++ src/main/java/subway/controller/SubwayController.java
@@ -1,19 +1,16 @@
 package subway.controller;
 
-import org.jgrapht.alg.shortestpath.DijkstraShortestPath;
-import org.jgrapht.graph.DefaultWeightedEdge;
-import org.jgrapht.graph.WeightedMultigraph;
 import subway.domain.Edge;
 import subway.domain.EdgeRepository;
 import subway.domain.Station;
 import subway.domain.StationRepository;
 
 import java.util.Arrays;
-import java.util.List;
 
 public class SubwayController {
-    private WeightedMultigraph<String, DefaultWeightedEdge> distanceGraph;
-    private WeightedMultigraph<String, DefaultWeightedEdge> timeGraph;
+    public void run() {
+
+    }
 
     public void initPrimary() {
         Arrays.asList("교대역", "강남역", "역삼역", "남부터미널역", "양재역", "양재시민의숲역", "매봉역")
@@ -27,33 +24,4 @@ public class SubwayController {
         EdgeRepository.addEdge(new Edge("양재역", "양재시민의숲역", 10, 3));
     }
 
-    public void initGraph() {
-        initDistanceGraph();
-        initTimeGraph();
-    }
-    
-    private void initDistanceGraph() {
-        this.distanceGraph = new WeightedMultigraph(DefaultWeightedEdge.class);
-        StationRepository.stations().forEach(station -> distanceGraph.addVertex(station.getName()));
-        EdgeRepository.edges().forEach(edge -> distanceGraph.setEdgeWeight(distanceGraph.addEdge(edge.getFrom(), edge.getTo()), edge.getDistance()));
-    }
-    
-    private void initTimeGraph() {
-        this.timeGraph = new WeightedMultigraph(DefaultWeightedEdge.class);
-        StationRepository.stations().forEach(station -> timeGraph.addVertex(station.getName()));
-        EdgeRepository.edges().forEach(edge -> timeGraph.setEdgeWeight(timeGraph.addEdge(edge.getFrom(), edge.getTo()), edge.getTime()));
-    }
-
-    public List<String> getShortestPathDistance(String from, String to) {
-        return getShortestPath(from, to, this.distanceGraph);
-    }
-    
-    public List<String> getShortestPathTime(String from, String to) {
-        return getShortestPath(from, to, this.timeGraph);
-    }
-
-    private List<String> getShortestPath(String from, String to, WeightedMultigraph<String, DefaultWeightedEdge> graph) {
-        DijkstraShortestPath dijkstraShortestPath = new DijkstraShortestPath(graph);
-        return dijkstraShortestPath.getPath(from, to).getVertexList();
-    }
 }